Move Triton launcher into `helion/runtime/triton/launcher.py`
Relocate the runtime symbols that Helion's *generated* Triton code depends on
at execution time out of the catch-all `helion/runtime/__init__.py` into a
dedicated `helion/runtime/triton/launcher.py`:

- `default_launcher` — runs a compiled `triton.jit` kernel
- `get_num_sm` — persistent-kernel grid size (host statement)
- `set_triton_allocator` (+ `_alloc_fn`) — TMA / tensor-descriptor scratch allocator

`helion.runtime` re-exports all three, so `helion.runtime.default_launcher`,
`helion.runtime.get_num_sm`, `helion.runtime.set_triton_allocator`, and the
generated `from helion.runtime import default_launcher` are unchanged.

This is a pure move with no behavior change — the module still reaches back into
a few `helion` helpers (`exc`, `settings`, `_utils`). Isolating these into one
small module is the first step toward making them dependency-free, so the
ahead-of-time precompiler can bulk-export exactly this file into a standalone
kernel.

Because `get_num_sm` now resolves its globals in the new module,
`test/test_runtime.py` is updated to patch
`helion.runtime.triton.launcher._module_is_pallas_interpret` (the moved
function's real lookup site) instead of `helion.runtime._module_is_pallas_interpret`.

Verified: `ruff`/`pyrefly` clean, `test_runtime` and `test_persistent_kernels`
pass, and the generated Triton for `examples/add.py` is byte-identical.

stack-info: PR: https://github.com/pytorch/helion/pull/3178, branch: AmesingFlank/stack/92
